+++ app/core/tests/test_get_users.py
@@ -4,7 +4,7 @@ from rest_framework.test import APIClient
 
 from constants import admin_project
 from constants import member_project
-from core.cru import Cru
+from core.api.cru import Cru
 from core.tests.utils.load_data import load_data
 from core.tests.utils.seed_constants import valerie_name
 from core.tests.utils.seed_constants import wally_name
