+++ app/core/tests/test_patch_users.py
@@ -6,8 +6,8 @@ from rest_framework.test import APIRequestFactory
 from rest_framework.test import force_authenticate
 
 from constants import admin_project
+from core.api.cru import Cru
 from core.api.views import UserViewSet
-from core.cru import Cru
 from core.tests.utils.load_data import load_data
 from core.tests.utils.seed_constants import garry_name
 from core.tests.utils.seed_constants import valerie_name
