+++ app/core/tests/unit_test/test_validate_fields_patchable_method.py
@@ -1,7 +1,7 @@
 import pytest
 from rest_framework.exceptions import ValidationError
 
-from core.permission_check import PermissionCheck
+from core.api.permission_check import PermissionCheck
 from core.tests.utils.load_data import load_data
 from core.tests.utils.seed_constants import garry_name
 from core.tests.utils.seed_constants import patti_name
